[MOD] presentation layer 분리
