+++ app/src/main/java/com/example/composestudy/presentation/MainActivity.kt
@@ -1,0 +1,14 @@
+package com.example.composestudy.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            SearchImageUI()
+        }
+    }
+}
